misc/HtmlTestSuite/results/test-0124.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Overflow: parsing valid scrollbar-gutter declarations</title>
<link href="mailto:felipeerias@gmail.com" rel="author" title="Felipe Erias Morandeira"/>
<link href="https://www.w3.org/TR/css-overflow-4/#scollbar-gutter-property" rel="help"/>
<meta content="Parsing valid scrollbar-gutter declarations" name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>

<script>

    test_valid_value("scrollbar-gutter", "auto");
    test_valid_value("scrollbar-gutter", "stable");
    test_valid_value("scrollbar-gutter", "stable both-edges");

    test_valid_value("scrollbar-gutter", "both-edges stable", "stable both-edges");

</script>
</head><body/></html>
